Update OperatingSystemsFunctions.txt
I was so absorbed in finding the appropriate information of operating systems, that I forgot to do this in the correct location. Although this is a bit mind boggling at first, its really quite simple to use. The instructions were easy to read, but the hardest part was to take a screenshot of my system build. After working on this a little bit, and listening to some videos, I feel more confident about learning more of Github's coding.