Update OperatingSystemsFunctions.txt
I just wanted to put another note before I find the correct set of commands to get my file correctly copied over. As easy as it was to use the websites items, it was hard to find the editing option. I kept going to the wrong page, and even on the right page you had to know where it was or you spent some time looking for it. It's very interesting to learn it just makes me a little nervous because its navigation on the website is a bit cluttered.